Add minimum upgrade base scenario for Horizon testing
soc8-horizon-min-for-upgrade.yml is a minimum upgrade scenario for testing
Horizon upgrade functionality. Use this to deploy Cloud 8, upgrade and then
check whether Horizon is still working.

Note that we will have to test Horizon again once all barclamps are in place
for the upgrade since some of them come with Horizon plugins of their own.