Move upgrade scripts to cloud8/
This is strictly for pre-merge testing:

cloud-mkphyscloud will need an extra parameter that allows
specifying a custom scenario directory for the upgrade
scenarios. This parameter cannot be added by specifying a
different automation repository branch upon launching the job
(it's too late at this point).

Hence this commit moves the upgrade scenarios to the cloud8/
directory temporarily. This allows them to be used without
specifying a custom scenario directory.